This adds ajax for group and personal playlists
The personal ajax call while creating a playlist is done.
It adds some divs for the ajax to find the playlist id in the
show page. Also adds a spinner that will help the user understand
things are loading in the background. The group is almost done,
just needs to have the divs added to the show page and it should be
good.